backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
markitdown[pdf]==0.1.5b1
PyPDF2==3.0.1
python-docx==1.1.2
langchain==0.3.18
langchain-core==0.3.35
langchain-groq==0.2.4
groq==0.16.0
python-dotenv==1.0.1
aiofiles==24.1.0
requests==2.32.3
jinja2==3.1.5
reportlab==4.2.5
html2text==2024.2.26